v1.17 XW 11-11-2019 fixed: caches (issue #15)

+++ src/org/jnode/fs/jfat/BootSector.java
@@ -358,7 +358,7 @@ public class BootSector {
         return BPB_Media;
     }
 
-    public long getSectorsPerFat() {
+    public int getSectorsPerFat() {
         return FATSz;
     }
